refs #11984 allow each transfer/mv step to handle stale files.

+++ leginon/batchtransfer.py
@@ -32,7 +32,7 @@ class RawTransfer(filetransfer.FileTransfer):
 		local drive removed to make room for more data collection.
 		'''
 		dirname,basename = os.path.split(os.path.abspath(dst_dir))
-		print 'transfer directory', src_dir, dst_dir
+		print('transfer directory', src_dir, dst_dir)
 		# get path of the session, e.g. /data/frames/joeuser/17nov06a
 		sessionpath = os.path.abspath(os.path.join(dirname,'..'))
 
@@ -82,7 +82,7 @@ class RawTransfer(filetransfer.FileTransfer):
 			for a in images_with_frame:
 				images[a['camera']['frames name']] = a
 			if images:
-				print s['name'],len(images)
+				print(s['name'],len(images))
 				sessions[s['name']] = images
 		return sessions	
 
@@ -99,10 +99,11 @@ class RawTransfer(filetransfer.FileTransfer):
 	def get_dst_head(self):
 		return self.getAndValidatePath('dest_path_head')
 
-	def getSessionFramePath(self, imdata):
+	def getSessionFramePath(self, session_name):
+		sessionq = leginon.leginondata.SessionData(name=session_name)
+		imdata = leginon.leginondata.AcquisitionImageData(session=sessionq).query(results=1)[0]
 		image_path = imdata['session']['image path']
 		frames_path = imdata['session']['frame path']
-
 		# use buffer frame path
 		# buffer server has access to both permanent and buffer frame path.
 		# Must be specific here.
@@ -198,22 +199,40 @@ class RawTransfer(filetransfer.FileTransfer):
 		goods = imkeys.intersection(filekeys)
 		return map((lambda x: {'image':image_map[x],'file':file_map[x]}), goods)
 
+	def validateDst(self, session_name, dst_head):
+		frames_path = self.getSessionFramePath(session_name)
+		if sys.platform != 'win32' and not frames_path.startswith(dst_head):
+			return False
+		else:
+			return frames_path
+
 	def runSession(self, session_name, image_file_maps, method, mode_str):
-		if not image_file_maps:
+		# 1. make sure destination is valid from params
+		dst_head = self.get_dst_head()
+		frames_path = self.validateDst(session_name, dst_head)
+		if frames_path is False:
 			return
+		# 2. if there are files in src_path, put them in a directory named by session_name
+		if image_file_maps:
+			self._moveToSrcSession(session_name, image_file_maps, frames_path)
+		# 3. if there are files in src_session_dir, transfer them
+		src_path = self.get_source_path()
+		src_session_dir = os.path.join(src_path,session_name)
+		dst_tmp_dir = os.path.join(frames_path,'tmp')
+		if os.path.isdir(src_session_dir):
+			self._transferSrcSessionToDstTmp(src_session_dir, dst_tmp_dir, method, mode_str)
+		# 4. check and move files in dst_tmp_dir/session_name to the final location.
+		self._moveDstTmpToFramesDir(session_name, dst_tmp_dir, frames_path)
+
+	def _moveToSrcSession(self, session_name, image_file_maps, frames_path):
 		first = list(image_file_maps)[0]
+		dst_suffix = first['file']['suffix']
 		imdata = first['image']
 		image_path = imdata['session']['image path']
 		uid, gid = self._getUidGid(image_path)
-		frames_path = self.getSessionFramePath(imdata)
-		dst_head = self.get_dst_head()
-		print dst_head
-		dst_suffix = first['file']['suffix']
-		if sys.platform != 'win32' and not frames_path.startswith(dst_head):
-			return
 		# move to a session temp directory on the same disk and then
 		# transfer since rsync is faster when doing the whole directory.
-		print 'name changing and moved to the src_tmp_dir'
+		print('name changing and moved to the src_tmp_dir')
 		src_tmp_dir = os.path.join(first['file']['dir'],session_name)
 		pyami.fileutil.mkdirs(src_tmp_dir)
 		for m in image_file_maps:
@@ -229,7 +248,7 @@ class RawTransfer(filetransfer.FileTransfer):
 					self.refcopy.setFrameDir(frames_path, gid, uid)
 					self.refcopy.run(imdata, dst_frame_name)
 				except:
-					print 'reference copying error. skip'
+					print('reference copying error. skip')
 			# de only  TODO: Not sure if the path is correct
 			info_path = os.path.join(src_tmp_frame_path,'info.txt')
 			if os.path.isfile(info_path):
@@ -240,18 +259,42 @@ class RawTransfer(filetransfer.FileTransfer):
 				xml_src_tmp_path = src_tmp_frame_path.replace('mrc','xml')
 				if os.path.exists(xml_src_path):
 					self.move(xm_src_path, xml_src_tmp_path)
+
+	def _transferSrcSessionToDstTmp(self, src_session_dir, dst_tmp_dir, method, mode_str):
+		files = os.listdir(src_session_dir)
+		first = False
+		for f in files:
+			try:
+				first = leginon.leginondata.AcquisitionImageData(filename=f.split('.')[0]).query(results=1)[0]
+				image_path = first['session']['image path']
+				break
+			except:
+				continue
+		if first is False:
+			# no frames files to process.  May have other junks
+			return
+		uid, gid = self._getUidGid(image_path)
 		# move the whole session together.
-		dst_tmp_dir = os.path.join(frames_path,'tmp')
-		print 'move dir to the dst_tmp_dir', dst_tmp_dir
+		print('move dir to the dst_tmp_dir', dst_tmp_dir)
 		# make temp dst_path
-		self.transfer_dir(src_tmp_dir, dst_tmp_dir, uid, gid, method, mode_str)
-		# the results are in dst_tmp_dire/session_name
-		#TODO: move *.* in dst_tmp_dir backout
+		pyami.fileutil.mkdirs(dst_tmp_dir)
+		self.transfer_dir(src_session_dir, dst_tmp_dir, uid, gid, method, mode_str)
+		# the results are in dst_tmp_dir/session_name
+		try:
+			os.rmdir(src_session_dir)
+		except Exception as e:
+			print('Error removing intermediate files: %s' % e)
+			
+
+	def _moveDstTmpToFramesDir(self, session_name, dst_tmp_dir,frames_path):
+		#move *.* in dst_tmp_dir backout
 		dst_tmp_session_dir = os.path.join(dst_tmp_dir, session_name)
+		if not os.path.isdir(dst_tmp_session_dir):
+			return
 		files = os.listdir(dst_tmp_session_dir)
 		for f in files:
 			tmp_f = os.path.join(dst_tmp_session_dir, f)
-			print 'final move from', tmp_f
+			print('final move from', tmp_f)
 			self.move(tmp_f, frames_path)
 		# clean up tmp_dir
 		os.rmdir(dst_tmp_session_dir)
@@ -262,18 +305,17 @@ class RawTransfer(filetransfer.FileTransfer):
 		method = self.params['method']
 		mode_str = self.params['mode_str']
 		src_path = self.get_source_path()
-		print 'Source path:  %s' % (src_path,)
+		print('Source path:  %s' % (src_path,))
 		file_map = self.getFrameFileMap(src_path)
 		dst_head = self.get_dst_head()
 		if dst_head:
-			print "Limit processing to destination frame path started with %s" % (dst_head)
+			print("Limit processing to destination frame path started with %s" % (dst_head))
 		session_maps = self.getSessionImageMaps()
 		for session_name in session_maps.keys():
-			print 'Iterating %s' % (session_name)
+			print('Iterating %s' % (session_name))
 			intersection_images = self.getIntersectionImages(session_maps[session_name],file_map)
 			self.runSession(session_name,intersection_images,method,mode_str)
 
 if __name__ == '__main__':
 		a = RawTransfer()
 		a.run()
-		#testRefCopy()
